On the request of tackat, change all floats to doubles.  We need the added precision when we are going to more detailed tiles.
An initial experiment of doing this automatically didn't come out too
well: All places except 3 disappeared, and the North Pole was placed
at the equator.  Instead I'm now doing it manually.  This commit
changes all floats in the classes Quaternion, GeoPoint and GeoPolygon,
but only as little as possible of the rest to make it compile.

From here on, I will take it step by step.


svn path=/trunk/KDE/kdeedu/marble/; revision=674069