Create the new class MarbleMap.
This is the first step in a refactoring that will allow Marble to be
used in situations where we cannot use a widget.  Examples include:
 - Plasmoids
 - KOffice Flake shapes.

The idea is to have an object -- MarbleMap -- that represents the map
and that can paint it on any QPainter that is given to it. This object
can be used almost exactly like MarbleWidget now, except that it is
not a widget.

The next step is to rewrite MarbleWidget to use MarbleMap and the
third step to polish the API a bit.


svn path=/trunk/KDE/kdeedu/marble/; revision=769785