Fix issue with integer overflow and some cleaning.
We have found a number of places where integer overflow occured due to
squaring of the radius.  The radius can potentially be very big (it
measures the current radius of the earth in pixels, which can get big
in high zoom levels) and squaring it can cause overflow.

This patch fixes all instances that I could find by using either 64
bit integers or doubles.


svn path=/trunk/KDE/kdeedu/marble/; revision=773138