* Move the MapQuality measure from ViewParams to ViewportParams.  * Some cleaning also snuck in.
Unfortunately I had to disable the geoshape since it uses the
installed headers and not the one in the source tree.  This breaks
compilation in a case like this where an fundamental API changes.
Enable it again once you have installed the new headers.  Beware,
though, that you will break compilation for everybody that haven't
installed the new headers yet.  That's the disadvantage of using
installed headers within a major source module of KDE.


svn path=/trunk/KDE/kdeedu/marble/; revision=814170