Introduce a read-only facade for internal Marble data.
This is usable e.g. for plugins that are allowed to access data, but
not modify it.  In the future it will perhaps expose data from more
classes than MarbleModel.

It is now up to tackat to use this in his stars plugin, that needs the
current time to show the starry background correctly.


svn path=/trunk/KDE/kdeedu/marble/; revision=835657