use the new threaded approach for loading kml data.
This will adapt the marbleWidget API to the idea of PlacemarkContainers who can be either from a file or from a QString.
This means as well that a container is unique and thus at the next loading it would not be replaced but rather the newly added container would be ignored.
To keep the binary compatibility for digikam, a container holding String data will now be removed before being reloaded. I am not sure but I guess this changes the behaviour of the marble widget a bit since this will restrict the shown placemarks to the files being in the current album. If this behaviour is not wanted, but rather all pictures should be shown on the map, we would need to adapt digikam to use a the album name as a key to make the containers distinguishable.
CCBUG:179307

svn path=/trunk/KDE/kdeedu/marble/; revision=905257