Remove the following methods from the public API:
void MarbleMap::setDownloadManager( HttpDownloadManager * );
void MarbleMap::setDownloadUrl( const QString & );
void MarbleMap::setDownloadUrl( const QUrl & );

void MarbleWidget::setDownloadManager( HttpDownloadManager * );
void MarbleWidget::setDownloadUrl( const QString & );
void MarbleWidget::setDownloadUrl( const QUrl & );

The only user of this API inside the KDE svn repository was
playground/base/plasma/applets/contacts/marblecontactswidget.cpp which is
changed already in svn revision 958083.

The download urls are stored in the dgml files for quite some time so
setDownloadUrl is really unnecessary and misleading because it will be
ignored. Was is being used are the values configured in the dgml.

As there is already a setDownloadManager method in MarbleModel, it does not
need to be in MarbleMap and MarbleWidget as well.

svn path=/trunk/KDE/kdeedu/marble/; revision=958290