This patch adds a context menu item "Remove Last Measure Point" for the right mouse button menu.
Author: "Michael G. Hansen" <mikeml2@pfna.de>

In addition, when the coordinate under the cursor is copied to the
clipboard using the left mouse button menu, it is also provided in the
clipboard as GPX and KML (can be retrieved by other applications by
requesting data in the respective mime-types). This makes it easy for
applications to understand coordinates independent of the plain-text
format.




svn path=/branches/marble/marble-gsoc-2009/; revision=979949