This patch adds basic infrastructure for host and usage specific download policies. It does not add features, only adds infrastructure and changes HttpDonwloadManager to use the new infrastructure.
Motivation:
===========
OpenStreetMap allows for bulk downloads (planned "Download region" feature)
only two concurrent connections. For browsing it is not a problem to have 10+
connections (like web browsers do).

So we need a way to distinguish between bulk downloading and browsing the map
per host/set of hosts.

What this patch contains:
=========================
- basic classes (DownloadQueueSet, DownloadPolicy)
- changes to HttpDownloadManager to use DownloadQueueSet internally
- one default download policy, which allows 20 concurrent connections
- change acceptJob to canAcceptJob and only create HttpJob if the job can be
  accepted

What has to be done:
====================
- agree on additional DGML elements/attributes for defining policies
- implement those DGML extensions
- make HttpJobs aware of their usage
...

svn path=/trunk/KDE/kdeedu/marble/; revision=1029978