Declare TextureTile::initJumpTables private.
initJumpTables is an operation which has to take place after the QImage
representing the original downloaded tile has changed, whether directly
after creation of a TextureTile object or after an update of an existing
TextureTile object because the content was expired.

One special thing about TextureTile::initJumpTables is that it has a
parameter (requestTileUpdate) specifying whether the signal tileUpdateDone
should be emitted.

There were two users of TextureTile::initJumpTables outside TextureTile.

1) TileLoader::loadTile, here TextureTile::initJumpTables was called with
   requestTileUpdate=false after TextureTile::loadDataset.
   The solution in this case was to call initJumpTables from within
   TextureTile::loadDataset.
2) MarbleModel::paintTile, here TextureTile::initJumpTables was called
   with requestTileUpdate, which is a parameter for paintTile also.
   MarbleModel::paintTile in turn is called from two places,
   a) TileLoader::loadTile, here with requestTileUpdate=false,
      so no change required
   b) TileLoader::reloadTile, here with requestTileUpdate=true,
      therefore we have to emit tileUpdateAvailable after the
      paintTile call.

svn path=/trunk/KDE/kdeedu/marble/; revision=1081105