Let TileLoader use HttpDownloadManager's downloadComplete( filename, id ) signal again.
This partly reverts the change for using the downloaded data directly w/o
loading it from the filesystem, because there is a problem with how the
cloud tiles are handled.

TileLoader::setLayer lets us specify for which map theme we want to load
tiles. The widget displays one map theme at a time, and when the users
switches to a different map theme, TileLoader::setLayer is called with the
new map theme as argument. The TileLoader internally manages a hash for fast
access to the tiles currently on display.

In MarbleModel::paintTile (with given tile), MergedLayerDecorator::paintTile
is called, the tile given is for example a tile of the bluemarble theme.
Then, in MergedLayerDecorator there is code to load the corresponding cloud
tile:
	TextureTile tile( TileId(...)); // here a empty TextureTile object is constructed

Then the tile's downloadTile signal is connected to MergedLayerDecorator's
downloadTile signal and then TextureTile::loadDataset is called.
In TextureTile::loadDataset the downloadTile signal is emitted (because the
tile was not yet locally available), which is connected to
HttpDownloadManager's addJob slot.
The tile is then downloaded and HttpDownloadManager emits the downloadComplete
signal, to which the TileLoader is connected. TileLoader then gets the data
and the TileId (which does not contain information about map themes, only
tile level, x and y), so the TileLoader updates the wrong tile, namely the
tile of the bluemarble theme.

svn path=/trunk/KDE/kdeedu/marble/; revision=1082697