DownloadRegionDialog::region: switch from tile coords to pixel coords.
When calculating tile coords, it is better to do it in pixel coords as long
as possible to not loose accuracy for no reason. With the current approach
of using the top level (of the pyramid) coords as reference this might not
give benefit, but once we switch to bottom level coords, there will be much
better accuracy.

svn path=/trunk/KDE/kdeedu/marble/; revision=1118971