Use the new monav daemon to retrieve routes. It will run out of the box, given that - MoNavD, the monav daemon, r225 or later is available in PATH - A compatible monav map lies in ~/.local/share/marble/maps/earth/monav. Version 0.0.1 is not compatible. You only need the contraction hierarchies and gps grid files. - A plugins.ini file with the content
[General]
router=Contraction Hierarchies
gpsLookup=GPS Grid

lies in the same dir.
signals.h (daemon socket interface) is taken from monav. It has a dual license for that purpose.
Note that the daemon uses little resources (map is not held in RAM) and can be left running in the background fine.

svn path=/trunk/KDE/kdeedu/marble/; revision=1172217