GeoDataPlacemark: Fix potential crash.
In the assignment operator of GeoDataPlacemarkPrivate m_geometry is deleted
without setting the pointer to 0 afterwards. In case the "other"
GeoDataPlacemarkPrivate does not have a geometry this might lead to a
double delete.

Besides that, the return statement might cause problems in case code
is added at the end of the method without carefully looking.

In addtition the geometry assignement is now exception-safe. Even if
exceptions are not used it seems nicer to use an existing wide spread
idiom ("first do the hard work, then assign").

svn path=/trunk/KDE/kdeedu/marble/; revision=1194384