backport of trunk-r1209186: introduce TileScalingTextureMapper
This patch introduces an optimized code path for the case that both the tile projection and the map projection are of type Mercator. In this very common case, the tiles can
be scaled rather than reprojected, promising a significant speedup, especially on the N900. Moreover, it seems to avoid crashes in high zoom levels on the N900, too.

REVIEW: 6109

svn path=/branches/KDE/4.6/kdeedu/marble/; revision=1210171