forward-port r1211240: turn GpFifo into a class and rename it to EmbossFifo
This patch resolves a small bug in the bumpmapping code where the gradient for the relief wasn't calculated properly.

It also turns the hidden union "GpFifo" in TextureColorizer.cpp into a proper FIFO class in order to make the code less fragile. The patch relies on very basic compiler
optimizations, relinguishing usage of a union.

I get ~3.5 fps on my machine before and after this patch is applied.

REVIEW: 6205

svn path=/branches/marble/marble-1.1/; revision=1215799