Fix frame item size calculation.
The total size must respect the border width if it exceeds the margin
(which is always 0 for our current plugins). Enforce a minimum padding
of 0, not more. If an automatic padding calculation is desired, it must
be done later since it depends on the border width and frame type and
both could change after calling setPadding. Still it makes sense to let
plugin authors decide on the value to set - e.g. to place items tightly
aligned to the border of a rounded rectangle.
(cherry picked from commit 9b7fc7519ca12e7764779a13c27fcd7bb5992910)