Start refactoring routing code: RoutingModel simplification.
The goal is to provide a routing API to users of libmarblewidget which
is easy to use for people knowing the QtMobility location classes.
Additionally the code needs simplification and a speedup at some
places.

Adds Maneuver, RouteSegment and Route classes. API written close to
the QtMobility 1.2 Location classes. Use these classes internally
in RoutingModel.