StackedTileLoader: also use expired tiles from cache (rather than throwing them away and loading them from disk again)
When an outdated tile is read from disk, new data is requested from the internet. As long as this data is not available, it does not make sense to re-read the tile from disk. Since the tile is updated in the cache as soon as the new data is available, there is neither a need for erasing expired tiles from the cache, nor to reload the tile from disk.