only update displayed tiles
A tile update is usually caused by interaction with the globe.
Since the process of updating a tile is rather expensive, it causes jerky behavior and should thus be minimized.
Moreover, if a tile has multiple layers, the effect of an update process may be experienced several times for a single tile.
On the other hand, if an update arrives for a displayed tile, the map should be updated.
As a result, only displayed tiles should be updated, while cached but non-displayed tiles should be removed from the cache.