improve perceived rendering performance: avoid loading of tiles from disk during rendering
== Problem ==
Whenever the bounding box of the map changed, tiles were potentially loaded from disk during rendering.
During e.g. panning, the disk access would cause small but annoying disruptions.
This effect was even worse when a tile consisted of multiple layers, where not only the disk access would cause disruptions, but also the tile merging.

== Solution ==
Rather than loading tiles from disk, Marble now uses a backing store, which is a tile cache that isn't modified during animations (e.g. panning).
If a tile that should be rendered is missing in the cache, the closest tile from the upper levels is used.
This is an ability which was introduced in the previous commit.

In order to enforce the presence of upper-level tiles, the tile pyramid of the visible bounding box (ranging from 0..<current level>) is loaded into the cache in still image mode.

== Results ==
Panning feels much smoother on the desktop as well as on the N900, especially when tiles consist of multiple layers.
Far animated voyages such as Berlin -> Berkeley feel much smoother (as long as the universe doesn't become visible).

== Possible improvements ==
* load more tiles surrounding the current bounding box (s.t. tiles of higher levels are less likely to appear)
* load and merge tiles in a separate thread