allow layers to query the current view context
This API change allows layers to be notified when Marble switches to e.g. the still context, where expensive operations may be performed w/o the user noticing.
In the animation context, layers may reduce expensive operations such that stalls are reduced.

The patch changes the API such that the map quality can only be set by setting still or animation context, where each context is associated with a MapQuality.