Revert "Fix another memleak due to wrong reference counting."
This reverts commit 9a1e0c398d5dda4fd9a45a271545914cda31623a.
Exchanges a crash for a memory leak. With the patch applied
saving a track leads to a crash some time after
PositionTracking.cpp:213. The deletion of the document
seems to delete some shared data as well. I'm not sure how
ownership in PositionTracking.cpp:206 is supposed to be.

CCMAIL: tgridel@free.fr