introduce ScanlineTextureMapperContext and clean up the texture mappers
A class named ScanlineTextureMapperContext is introduced, which holds the state for performing scanline texture mapping. It was
factored out of the AbstractScanlineTextureMapper class, which is now named "TextureMapperInterface". Fresh
ScanlineTextureMapperContexts are created upon every rendering action, which is cleaner than the former approach of reinitializing all
rendering attributes in AbstractScanlineTextureMapper::setRadius(). Finally, the patch constifies many variables in the texture mappers
and removes some unneeded code.