Offline address search database generator.
Extracts placemarks (addresses and POIs) from OpenStreetMap .osm xml
files and .pbf binary files. Stores them in a sqlite database file.

Code LGPL by me except that in the pbf/ directory, see the origins
and licenses in the respective file headers and/or the README.txt in
that directory.

This code is not part of the Marble library or any of the Marble
executables. It is not useful for end-users. It is not build by cmake.
Hence it is not affected by the various freezes.