Reduce memory usage by about 75%.
Let the pbf parser do a three-pass scan of the file to be able to know
which ways and nodes are simple (no address/poi data) and are not
referenced by any administrative boundary and thus can be discarded.
Distinguish between nodes and coordinates (simple nodes). Use hashes
instead of maps for faster lookups. Free data not needed anymore as
early as possible. The memory needed to process a .pbf file is now
around three times the size of the file (varies depending on the data).