Memory optimization for people using a large number of offline maps.
Country bounding box polygons are rather big for some countries in the
new offline routing maps. When all countries are installed in Marble,
they sum up to ~75 MB which is too much for the mobile version. Discard
those larger than ~40 kB and only store the rectangular bbox as an
approximation. Work around the to be expected false positives by
querying other matching country maps if the first guess was wrong.