move MarbleWidgetPopupMenu from the input handler to MarbleWidget
The input handler should ideally only filter inputs and trigger actions.
The actions should either be implemented in the MarbleWidget or in plugins, but -- ideally -- not in the input handler.

Showing a popup menu is the implementation of an action.
Therefore, the popup menu should not be part of the input handler.
Moreover, the popup menu should be accessible by plugins for adding their actions, such that the best place for it seems to be the MarbleWidget.