determine visibility of placemarks in MarbleMap rather than in PlacemarkLayout
The first motivtion for this patch comes from having PlacemarkLayout inherit from LayerInterface in the future.
This can only be achieved if PlacemarkLayout doesn't access ViewParams and GeoScene* classes.

The second motivation comes from making PlacemarkLayout more self-contained, i.e. not depending on GeoScene* classes.
Visibility attributes are not determined from a pointer to GeoSceneDocument any more, which may be zero.
Instead, visibilities are determined from members of PlacemarkLayout, which should make the latter less potentially fragile.
Moreover, the grand plan is to have QActions in the legend one day, which are connected to setShowFoo( bool visible ) slots of PlacemarkLayout.

REVIEW: 101842