move atmosphere into the "SURFACE" layer
Being currently in the "BEHIND_TARGET" layer, the atmosphere is currently considered a part of the background (layers "STARS" and "BEHIND_TARGET").
However, it currently gets rendered regardless of what MarbleMap::showBackground() returns.
Therefore, in order to keep current behavior once all layers are ported to the LayerInterface, we move it to the "SURFACE" layer and assign it a negative zValue().

Note that this commit won't have an effect until AtmosphereLayer is rendered by the LayerManager (which it is currently not).

RevBy: tackat