HttpDownloadManager: neither take ownership of nor expose StoragePolicy
The StoragePolicy is accessed outside of HttpDownloadManager, where methods are accessed that are only available in subclasses.
Since it was unsafe to keep a pointer to the concrete StoragePolicy around, downcasts from HttpManager's storagePolicy() were needed to access these methods.

By not having HttpDownloadManager take ownership, it is now safe to keep a concrete object of StoragePolicy around.
As a result, casts are unnecessary.