when latitude exceeds valid range for Mercator, clip to maxLat/minLat instead of reverting to the last known position
Reverting to the last known position caused some weired behavior when approaching the upper/lower borders of a Mercator-projected map:
When approaching the borders rather slow, one could move closer to the border compared to when approaching fast.

This patch allows to move to maxLat/minLat of a Mercator-projected map, regardless of how fast the borders are approached.