Install decl. plugins in CMAKE_INSTALL_PREFIX unless /usr/local is used
Declarative plugins are loaded by Qt from cwd, QLibraryInfo::ImportsPath
and QML_IMPORT_PATH. Setting up the latter is inconvenient for users,
so we try to avoid it:
- When CMAKE_INSTALL_PREFIX is in line with QML_IMPORTS_PATH, be happy
- When CMAKE_INSTALL_PREFIX is /usr/local (default), install to
QML_IMPORTS_PATH (cmake, something like /usr/lib/qt4/imports) even if
that is out of CMAKE_INSTALL_PREFIX and be quiet about it to avoid
confusing people reading the compile output.
- When CMAKE_INSTALL_PREFIX is different (e.g. in $HOME), respect it
when installing to avoid permission problems and instruct the user to
setup the QML_IMPORT_PATH environment variable.