Placemarks which are not on or above Earth in the view are now clickable
Additionaly, the coordinates used in both the left mouse button and right
mouse buttons are those of the first placemark under the cursor if it exists.
Before, the coordinates were determined by doing an inverse projection of the
cursor position on the globe, but this doesn't make sense for placemarks which
are not on the globe.