GoToDialog: keep current zoom value when centering on a placemark
When selecing a placemark in the GoToDialog, Marble used to change the zoom to a rather high value.
So when searching for a city, e.g. Berlin, a very small area was shown, even when whole Berlin was visible before.

Leaving the zoom value untouched when selecting a placemark in the GoToDialog, the user is able to select the desired zoom value upfront, and navigate to different search results w/o having to restore the desired zoom value.

RevBy: Earthwings

REVIEW: 102979