rework lmb and rmb menus
The lmb as well as the rmb menu used to offer actions regarding the current geo position. In particular, the lmb menu offers "Copy
Coordinates" and "Address Details", while the rmb menu offers "Add Bookmark" and "Set Home Location". As a result, neither of the
menus have precise duties, creating ambiguities, which this patch addresses as described below.

The lmb menu is now only responsible for dealing with links of geo features (in analogy to a web browser): When the mouse cursor
hovers above a feature (indicated by turning into an pointing hand), then a lmb click will reveal the lmb menu. If the mouse is not a
pointing hand, no lmb menu will appear when clicking the left mouse button.

No other actions than geo feature actions are part of the lmb menu. In particular, the unimplemented "planet" action is removed
entirely, whereas the remaining actions basically went into the rmb menu.

The rmb menu's purpose is to provide context actions for the respective geo position. It can be triggered anywhere and anytime
above the map (mouse cursor above a valid geo position).
Since the "Add Bookmark" and even more so the "Set Home Location" actions are believed to be very rarely used, they are removed in the
desktop version (but they are still available from the application menu). The mobile version keeps the "Add Bookmark" action, but
the "Set Home Location" is removed as well.
The "Copy Coordinates" and "Address Details" actions (removed from the lmb menu) are both moved to the rmb menu in the desktop version.
The mobile version gains the "Address Details" action.

BUG: 282394
REVIEW: 103053
GUI