really snap to optimal radius
We used to add +2 to the optimal radius in the MarbleWidget in order to make TextureLayer select a sharp tile level.
If we had passed the optimal radius to the TextureLayer, it would sometimes choose the next less precise tile level, resulting in blurred maps.

With this patch, the rounding errors are worked around by making the TextureLayer choose the sharper tile level a tiny bit earlier.