Improve where the streets have no names.
Generate turn instructions when entering an unnamed street. Note that
currently this will generate a number of false positives (instructions
somewhere in the middle of an unnamed street) because monav often
reports waypoints to be intersections when they are not. Using the same
heuristic as the monav client to avoid this (suppress instructions when
going into mostly straight direction). Appararently this is not
perfect.
BUG: 279424
FIXED-IN: 1.2.90