Refactoring: Factor out some stuff into new class NwwMapImage.
The methods nwwPixel, nwwTile, lonRadToNwwPixelX and latRadToNwwPixelY
as well as corresponding data members are now in class NwwMapImage.

Besides tileId is now a method and some issues with wrong detection
of missing tiles - where pixel coordinates instead of tile coordinates
were used to compose the tileId - in nwwPixel are fixed.

Only compile-tested.