mapreproject: Prepare for support of multiple input maps.
Instead of boost program_options now getopt_long and getsubopt are used
as boost program_options does not seem to support suboptions.

Other changes in this commit are the introduction of a new class
ReadOnlyMapDefinition which holds the data of a --input argument
and is used to create objects of ReadOnlyMapImage derived classes.

OsmTileClusterRenderer was changed to support multiple map sources,
the NwwMapImage member was replace by a vector of ReadOnlyMapImage
pointers.