Work around a crash triggered by tile creation on map theme switching.
When tiles need to be created during a map theme switch, the file
system watcher picks up the changes on the disk and updates the map
theme models. This leads to invalidated listview delegates. If the map
theme switch was triggered by such a delegate, a crash occurs. This
seems to be a general limitation in QML (A view delegate causing its
own invalidation leads to a segfault).
The workaround is not too great: A timer outside the scope of the
delegate triggers the change now.
(cherry picked from commit cd10ddf0c6f000e93175eb4a31e3d02ce551f949)