Fix the encoding errors in decoding the postal JSON data
The data is encoded in UTF-8, so just apply the proper decoding before
parsing the JSON data. Previously, this code decoded it as Latin 1,
parsed the JSON, then for the place names that could include non-ASCII
characters, it re-encoded as Latin 1 and tried to decode as UTF-8, via
std::string!