introduce DialogConfigurationInterface for plugins that wish to provide a QWidget-based configuration dialog
We'll use a qobject_cast now in order to determine whether a plugin offers a configuration dialog or not.
This is more efficient in terms of time and memory than trying to create a dialog and check whether the returned pointer is non-zero.

For QML-based platforms such as Mer or MeeGo, this approach allows to compile the plugins w/o QWidget support.