move guiString() into abstract classes, move initialize() and isInitialized() into interfaces
guiString() returns meta-knowledge about a plugin. newInstance(), which is declared in the abstract plugin classes, represents meta-knowledge about a plugin, too. Therefore, it makes sense to move guiString() into the abstract classes.

The interface classes, on the other hand, declare a protocol for certain types of features. No meta-knowledge should be required for their use, so guiString() should not be declared there. initialize() and isInitialized(), however, are part of the protocol, so they should be part of the interface.