Remove the mutex attribute.
Geodata objects are not thread-safe by itself, so the problem this
part of the code was solving should be handled elsewhere. From my
tests everything works still fine without the mutex.

This should solve the problem of creating a huge amount of handles
on Windows, reported by Konrad Enzensberger - thanks for reporting
and tracking it down!.

Please test.
CCMAIL: marble-devel@kde.org