factor out ElevationProfileMarker from ElevationProfileFloatItem
The elevation profile float item, a 2D object, used to render the altitude marker in 3D space.
This makes it impossible to treat the float item as a real 2D object.
By creating a separate item for the altitude marker, this issue can be resolved.

Being able to treat the float item as a real 2D object allows for easier caching for the upcoming OpenGL mode and it allows to introduce a float item layer.