update altitude marker just before it gets rendered in renderOnMap()
According to their respective render positions, the altitude marker is rendered before the float item.
Before this patch, the altitude marker's state was only updated when the float item got rendered, meaning it was rendered with an old state.
As a workaround, a second repaint request was generated.
With this patch (and with the previous one), the altitude marker's state is now updated just before it gets rendered, avoiding the need for a second repaint request.