Fix drawing of the default position marker.
The fix is pretty similar to the one in cb6992d3: Calculating the
current rotation based on the screen coordinates of the last two
positions is sensitive to numerical instabilities, so use the
direction provided by the plugin.

Note that this commit and cb6992d3 introduce a less general handling
of map rotations: Only north and south pointing upwards are supported.
Currently that's the case in Marble, but might change in the future
(e.g. OpenGL mode) and needs to be respected in the two rotate() calls
then.
CCMAIL: bbeschow@cs.tu-berlin.de