Respect updates in http://wiki.openstreetmap.org/wiki/Tile_Usage_Policy
The maximum tile zoom level for mass download is now restricted to 16
(prior 18). Also "fixes" the effect that downloads of high zoom levels
(17, 18) take considerably more time than zoom levels up to 16 -- which
happens because they are rendered on demand by the OSM servers.
Note that zoom level 17 and 18 can of course still be viewed in the map
and are cached like other zoom levels. The change only affects the
"Download Region" dialog.
CCBUG: 290671