Handle viewport IDL crossing when calculating visible items.
BUG: 298498
Note that commit 635dc5db mentioned in bug 298498#c0 did not cause the
bug itself, but only revealed it: Previously the bounding box was only
set in setLineString(), but not in the ctor which took the linestring
as well. Hence when using the ctor alone (which is what we do), the
bbox was not set to the correct one, but an empty one. Now the default
ctor of the bbox returns a sphere sized bbox, hence seemingly making it
work (hiding both the bbox calculation bug as well as the negative
performance implications). To me that's another argument to change the
default ctor of GeoDataLatLon*Box.
CCBUG: 299959