Online services accept a QML delegate for rendering/interaction.
A QML component can be registered to take over the rendering of the
layer. The C++ part takes care of the lifetime and positioning of the
items, QML is responsible for rendering and interaction. All item
properties (read: Q_PROPERTY ones) are available in the QML context
of each item under their names (context object).

Note that due to the separate rendering, the QML layer is painted on
top of other items rendered by C++/GeoPainter, e.g. float items.