PlacemarkPainter: turn into layer and rename to PlacemarkLayer; reverse composition with PlacemarkLayout
PlacemarkLayout used to be both responsible for laying out placemarks and painting them.
After having determined the layout, it realized painting by delegating the acutal work to PlacemarkPainter.

By reversing composition, PlacemarkPainter would ask PlacemarkLayout to update the layout before it would paint the placemarks itself.
That way, PlacemarkPainter becomes a layer and PlacemarkLayout would be the layer's model, which is a common pattern in Marble.