Support layers where both data and rendering are done from QML
Think of it as a QML (List,Grid)View where the layout is defined
by the geo position of the items in the model (instead of their
order transferred to some grid).
Based on a patch by Anton Chernov - thanks! Please see
examples/qml/data-layers/ how to use it. Basically you append one
or more DataLayer items to the dataLayers property of MarbleWidget.
Each DataLayer has a model property (use e.g. a ListModel), where
the lon, lat properties of elements define their geoposition. Optionally
the alt and identifier properties can be set. Those and all additional
properties are available in the qml context of each delegate (assigned
to the DataLayer via the delegate property). The delegate can be any
QML component.