Provide a fallback "icon", fix width sometimes wrong.
The remote icon does not always exist or can fail to load. Use a similar
looking rectangle as a fallback. Change animations not to set the
width property on their own, which does not always work and results in
wrong item sizes then. Instead calculate the width directly in the
property and only animate the result (via a Behavior).