Items can set themselves sticky.
When items change their size, collisions with other items can happen
that result in one of the items being excluded from rendering. When
the size change is the result of interaction, the size-changing item
shall not be excluded from rendering since it is the one the user
interacts with. To tell the renderer about such situations, items
can now use the sticky property that affects the rendering priority:
sticky > favorite > custom item priority.