faster background painting for GeoLineStringGraphicsItems
Line strings with a dash pattern and a background color (e.g. railways) were painted twice: First, the line string was painted using background color and then was painted again with a dash pattern using a transparent (!) background. Painting can be done in one step if an opaque background is used for the "spaces" in the dash pattern.

This approach also respects the cap style.