Introduce FloatItemsLayer, a dedicated layer for rendering screen-positioned, 2D float items
AbstractFloatItems are screen-positioned items that live in a 2D world. However, by overriding LayerInterface::render(), it was possible for float items to render in 3D space.

This patch avoids Marble calling 3D rendering API on AbstractFloatItems. As a result, the items are always rendered on the "FLOAT_ITEM" layer and 3D paintings aren't possible any longer for float items.