Fix folder doubling upon adding new folders. Use document as parent.
The combo box presents a flattened list of the folder hierarchy. Using
the currently selected folder as the parent of a newly added folder is
little intuitive and prevents to create top level folders. Instead now
top level folders are always created. The user can still create a
hierarchy in the bookmark manager if he wishes (where a tree view is
used, making the tree approach clear).