Extension for the .osm file parser.
Extension for the .osm file parser. Relations can be built with
different ways (for example for describing a building with a hole in it)
tagged as members.

So with this diff, <member> tag is parsed and ways are stored for then a
relation to have them available if necessary.

Moreover a relation can also contain another relation, so a relation
factory has also been created to store them.
REVIEW: 106007