API change: remove unused setPolygon() and setLinearRing() from GeoPolygonGraphicsItem
Being a lightweight rendering wrapper for GeoDataPolygons/GeoDataLinearRings, it does not make much sense to reuse a GeoPolygonGraphicsItem by assigning it another polygon or linear ring.
So if a different GeoDataPolygon/GeoDataLinearRing should be rendered, the old GeoPolygonGraphicsItem has to be removed and a new one has to be created and added to the scene.
This ultimately allows the renderer to know when the scene changes.