Add support for TileMapServer storageLayout in DGML files
Add support for TileMapServer storageLayout in DGML files (it is similar
to OpenStreetMap mode, but y coordinates start at the bottom of the map
rather than at the top).
Changes include a new TmsServerLayout class to transfom the y coordinate
and modifications in  GeoSceneTexture::relativeTileFileName to take into
account the new TmsServerLayout (tiles are stored locally with the same
filename as it would have in the remote server, allowing the use of
locally stored maps without changing the filenames).
As a side note, running "marble --tile-id" will show the original marble
coordinates (y starting at top), not the TMS ones, which can be
confusing. Moreover, two textures can be mixed with different
modes (i.e. OpenStreetMap and TileMapService), and only the OSM tile-id
will be shown.
REVIEW: 104042